refactor: move `HomeDir()` to `fsext` package
